chore: Optimize garbage collection for `Topology` instance (#33)
Extract crawl iteration run to a separate method. This way the local
`topology` instance is cleaned up from memory immediately after the
crawl.

Before this PR the `topology` instance was kept in memory until the
start of the next iteration. This happens because the method is async,
in a non-async method the `topology` would have been cleaned-up earlier
(already when it execution exists the `try-catch` scope as it is a local
variable there).

## Other changes

Also simplified handling of `subcriberGate`. There is no need to pass it
as an argument as it is a field of Crawler.